+++ source/TestCommon/documents/release-notes/release-notes.md
@@ -2,9 +2,15 @@
 
 ## Version 6.3.0
 
-- On class `IntegrationTestConfiguration`:
+- Refactored class `IntegrationTestConfiguration`:
     - Deleted obsolete property `ApplicationInsightsInstrumentationKey`.
     - Marked property `ServiceBusConnectionString` as obsolete.
+    - Added property `ServiceBusFullyQualifiedNamespace`.
+- Refactored class `ServiceBusResourceProvider`:
+    - Marked constructor with 'ConnectionString' parameter as obsolete.
+    - Marked property 'ConnectionString' as obsolete.
+    - Added constructor with 'FullyQualifiedNamespace' parameter.
+    - Added property 'FullyQualifiedNamespace'.
 
 ## Version 6.2.0